Add pagination to the /inventory/all API endpoint
Now the /inventory/all endpoint can be queried with the following
optional parameters as querystrings:

 - pagesize: Amount of items to be shown per page
 - page: The number of the page to be shown (starts at 1)

The endpoint can be queried up to an unlimited page number, but the
response will be an empty array if no items are left to be shown.

Despite both parameters being optional, either both or none of them
should be used. Querying the endpoint with only one of them will produce
an HTTP 400 error.